% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatReads.R
\name{StatRead}
\alias{StatRead}
\title{Quality Statistics}
\usage{
StatRead(samplename, files, type = NULL)
}
\arguments{
\item{samplename}{character, the sample names}

\item{files}{character, the path of fastq file.}

\item{type}{character, the prefix of output, eg:rawR1 or rawR2}
}
\description{
Quality Statistics
}
\author{
Shuangbin Xu
}
